+++ lib/ecto_shorts/query_builder.ex
@@ -1,20 +1,47 @@
 defmodule EctoShorts.QueryBuilder do
-  @moduledoc "Behaviour for query building from filter tuples"
+  @moduledoc """
+  Specifies the query builder API required from adapters.
 
-  @type filter_tuple :: {filter_type :: atom, value :: any}
-  @type accumulator_query :: Ecto.Query.t
+  An adapter receives the `Ecto.Queryable` (for example your ecto schema)
+  the query is built on, the filter key and value, and the accumulator
+  query that the filter is added to.
+  """
 
-  @doc "Adds to accumulator query with filter_type and value"
-  @callback create_schema_filter(filter_tuple, accumulator_query) :: Ecto.Query.t
+  @type adapter :: module()
+  @type filter_key :: atom()
+  @type filter_value :: any()
+  @type query :: Ecto.Query.t()
+  @type queryable :: Ecto.Queryable.t()
 
-  @spec create_schema_filter(module, filter_tuple, accumulator_query) :: Ecto.Query.t
-  def create_schema_filter(builder, filter_tuple, query) do
-    builder.create_schema_filter(filter_tuple, query)
+  @doc "Adds to the accumulator query with the filter key and value"
+  @callback build_query(
+    queryable :: queryable(),
+    filter_key :: filter_key(),
+    filter_value :: filter_value(),
+    query :: query()
+  ) :: query()
+
+  @doc """
+  Invokes `c:build_query/4` on the given adapter.
+
+  ### Examples
+
+      iex> EctoShorts.QueryBuilder.build_query(
+      ...>   EctoShorts.QueryBuilder.Common,
+      ...>   EctoShorts.Support.Schemas.Post,
+      ...>   :first,
+      ...>   10,
+      ...>   EctoShorts.Support.Schemas.Post
+      ...> )
+  """
+  @spec build_query(
+    adapter :: adapter(),
+    queryable :: queryable(),
+    filter_key :: filter_key(),
+    filter_value :: filter_value(),
+    query :: query()
+  ) :: query()
+  def build_query(adapter, queryable, filter_key, filter_value, query) do
+    adapter.build_query(queryable, filter_key, filter_value, query)
   end
-
-  @spec query_schema(Ecto.Queryable.t) :: Ecto.Queryable.t()
-  @doc "Pulls the schema from a query"
-  def query_schema(%{from: %{source: {_, schema}}}), do: query_schema(schema)
-  def query_schema(%{from: %{query: %{from: {_, schema}}}}), do: schema
-  def query_schema(query), do: query
 end