+++ test/ecto_shorts/query_builder/common_test.exs
@@ -3,7 +3,9 @@ defmodule EctoShorts.QueryBuilder.CommonTest do
   doctest EctoShorts.QueryBuilder.Common
 
   alias EctoShorts.QueryBuilder.Common
-  alias EctoShorts.Support.Schemas.Comment
+  alias EctoShorts.Support.Schemas.{Comment, Post}
+
+  require Ecto.Query
 
   describe "filters: " do
     test "returns expected list" do
@@ -24,11 +26,43 @@ defmodule EctoShorts.QueryBuilder.CommonTest do
     end
   end
 
-  describe "create_schema_filters: " do
-    test "returns query without changes when passed {:search, term()}" do
+  describe "build_query: " do
+    test "returns query without changes when searching a schema that does not define by_search/2" do
       expected_query = Comment
 
-      assert ^expected_query = Common.create_schema_filter({:search, %{id: 1}}, expected_query)
+      assert ^expected_query = Common.build_query(Comment, :search, %{id: 1}, expected_query)
+    end
+
+    test "returns query built by by_search/2 of the given schema" do
+      query = Common.build_query(Post, :search, %{id: 1}, Ecto.Query.from(Post))
+
+      assert %Ecto.Query{
+        from: %Ecto.Query.FromExpr{
+          source: {"posts", EctoShorts.Support.Schemas.Post}
+        },
+        wheres: [
+          %Ecto.Query.BooleanExpr{
+            expr: {:==, [], [{{:., [], [{:&, [], [0]}, :id]}, [], []}, {:^, [], [0]}]},
+            op: :and,
+            params: [{1, {0, :id}}],
+            subqueries: []
+          }
+        ]
+      } = query
+    end
+
+    test "returns query with limit when passed :first" do
+      query = Common.build_query(Post, :first, 10, Post)
+
+      assert %Ecto.Query{
+        from: %Ecto.Query.FromExpr{
+          source: {"posts", EctoShorts.Support.Schemas.Post}
+        },
+        limit: %Ecto.Query.LimitExpr{
+          expr: {:^, [], [0]},
+          params: [{10, :integer}]
+        }
+      } = query
     end
   end
 end